+++ src/frontend/back_propagation_numerical.cc
@@ -15,7 +15,7 @@ using namespace Eigen;
 constexpr size_t batch_size = 1;
 constexpr size_t input_size = 4;
 
-void program_body( const unsigned int num_iterations, const float epsilon )
+void program_body( const unsigned int num_iterations, const double epsilon )
 {
   /* remove limit on stack size */
   const rlimit limits { RLIM_INFINITY, RLIM_INFINITY };
@@ -31,7 +31,7 @@ void program_body( const unsigned int num_iterations, const float epsilon )
 
   srand( 10 );
   /* initialize inputs */
-  Matrix<float, batch_size, input_size> input = Matrix<float, batch_size, input_size>::Random();
+  Matrix<double, batch_size, input_size> input = Matrix<double, batch_size, input_size>::Random();
 
   nn->apply( input );
   const IOFormat CleanFmt( 4, 0, ", ", "\n", "[", "]" );
@@ -52,10 +52,10 @@ void program_body( const unsigned int num_iterations, const float epsilon )
          << endl;
 
     unsigned int numParams = nn->getNumParams( layerNum );
-    vector<float> gradients( numParams, 0 );
+    vector<double> gradients( numParams, 0 );
 
     for ( unsigned int paramNum = 0; paramNum < numParams; paramNum++ ) {
-      float gradient = nn->calculateNumericalGradient( input, layerNum, paramNum, epsilon );
+      double gradient = nn->calculateNumericalGradient( input, layerNum, paramNum, epsilon );
       gradients[paramNum] = gradient;
     }
 
@@ -82,7 +82,7 @@ void program_body( const unsigned int num_iterations, const float epsilon )
   // }
 
   // for ( unsigned int i = 0; i < num_iterations; i++ ) {
-  //   float gradient = nn->calculateNumericalGradient( input, gradientLayerNums[i], gradientWeightNums[i], epsilon
+  //   double gradient = nn->calculateNumericalGradient( input, gradientLayerNums[i], gradientWeightNums[i], epsilon
   //   ); cout << "Iteration " << i << ":\n"
   //        << "  Layer " << gradientLayerNums[i] << "\n"
   //        << "  Weight " << gradientWeightNums[i] << "\n"
